AI/k8s_monitor_agentAI/mcp_project/mcp_eks/requirements.txt
google-generativeai>=0.8.0
mcp>=1.0.0
slack-bolt>=1.18.0
slack-sdk>=3.27.0
httpx>=0.27.0
python-dotenv>=1.0.0
cffi
aiohttp
google-genai
aiohttp